refactor(user.controller): merge userController into user.controller
